refactor: [#661] move HTTP Tracker Client mod
