+++ src/console/clients/mod.rs
@@ -1,2 +1,3 @@
 //! Console clients.
 pub mod checker;
+pub mod http;
